+++ src/builtin/widgets/scroll/widget.rs
@@ -42,7 +42,7 @@ unimplemented
     - https://stackoverflow.com/a/7224899
 - macos-style 'jump one page on scrollbar press'
     - needs animation framework overhaul or bespoke solution
-        - bespoke solution likely best: need to also support pagination via mouse scroll events
+        - bespoke solution likely best: need to also support pagination via mouse scroll events and gamepad/controller inputs
     - The 'animate to next page' setting does the following (if you press on the bar and not on the slider handle)
         1. On press, animate one page in the direction of the cursor.
         2. Delay
@@ -68,12 +68,15 @@ be reflected in handle size until the next tick (an off-by-1 glitch)
 
 /*
 Demos
-Note: use buttons to add/remove content from scroll areas
+Note: use buttons to add/remove rows from scroll areas
+    - for 'expand horizontal', make the 'content blob entity' expand in width on press
+Note: add name to all scroll areas
+    - add text with arrow that points at scrollbar for the 'entire window' scrollbar
 
-- Entire window is a scroll area using 'show scroll bar when scrolling'
+- Entire window is a scroll area using Firefox style
     - Need the 'fold' between elements to overlap w/ bottom of window to imply scrolling is possible.
 
-- Basic
+- Basic vertical
 
 - Scrollbars as overlays
     - invisible scrollbar, visible but semi-transparent handle
@@ -83,19 +86,21 @@ Note: use buttons to add/remove content from scroll areas
         - Scroll area and vertical bar in row together
         - Horizontal bar
 
-- Scrollbar with FocusPolicy::Block and Pickable { should_block_lower: true, is_hoverable: true }
+- ScrollArea with FocusPolicy::Block and Picking::Sink
     - layered scrolling does not occur
 
-- Styling
+- SublimeText (bi-directional)
     - Shadow behind vertical scrollbar when there is a horizontal bar.
         - Add shim node behind bar with Static<NodeShadow>{state:["HorizontalScroll"]}, in control group with scroll base
     - Show/hide horizontal bar when there is scrollable content
         - Add shim node with Static<DisplayControl>{value:Hide} and {state:["HorizontalScroll"] value:Display}
     - Invisible/visible vertical bar when there is scrollable content (manual 'scrollbar-gutter')
+
+- Firefox (for entire scroll area)
     - Show scroll bar when scrolling
         - Update
             - Update scroll timers
-                - when timer ends, remove 'is scrolling' state
+                - when timer ends, remove 'is scrolling' and 'hover activated' states
         - ScrollBase
             - On MouseScroll event
                 - if no scrolling timer
@@ -104,26 +109,29 @@ Note: use buttons to add/remove content from scroll areas
                 - if scrolling timer
                     - refresh timer
         - ScrollBar
+            - Initial bar and handle opacity: zero
             - On pointer hover or press while in state 'is scrolling'
                 - refresh scrolling timer
-            - On enter 'is scrolling' state
-                - animate opacity of bar handle to target
-                - static
-                    - add Enabled state (enables interactions) -> tie animations to 'Enabled'
-                    - add picking behavior: Pickable { should_block_lower: false, is_hoverable: true }
-                        - don't block lower so scrolls can reach the scroll base (scrolls don't use picking, but we use
-                        picking hit stacks to decide where scroll amounts should go)
+                - add 'hover activated' state
             - On enter 'not scrolling' state
-                - animate opacity of bar handle to zero after delay
+                - animate opacity of bar and handle to zero after delay
                 - static
                     - add Disabled state (disables interactions)
-                    - add picking behavior: Pickable { should_block_lower: false, is_hoverable: false }
-            - Slider bar
-                - Set initial opacity to zero with plain loadable
-                - In no state: idle value of slider bar is '0.0 alpha', with enter animation
-                - on hover
-                    - 
-                - In state 'is scrolling': hover value of slider bar is '1.0 alpha', animated with zero duration and delay so the
-                attribute overrides the idle attribute
+                    - add picking behavior: Picking::Ignore
+                    - set width to 'small'
+            - On enter 'is scrolling' state
+                - animate opacity of bar and handle to target
+                - static
+                    - add Enabled state (enables interactions) -> tie animations to 'Enabled'
+                    - add picking behavior: Picking::Pass
+                        - don't block lower so scrolls can reach the scroll base (scrolls don't use picking, but we use
+                        picking hit stacks to decide where scroll amounts should go)
+                    - add Visibility::Hidden to bar (opacity animates 'hidden')
+                        - need Visibility::Visible on handle to avoid inheritance
+            - On enter 'hover activated' state
+                - static
+                    - add Visibility::Show
+                    - set width to 'large'
+                        - note: need to set radius of handle to 50% of bar width
 
 */